refactor(reason-bindings): use scoped filenames convention
Namespacing is quite important in Reason/Ocaml because all modules are
available in scope and there might be a clashing naming problem.
Bucklescript has an option to automatically namespace the modules, based
on the package name.
https://bucklescript.github.io/docs/en/build-configuration.html#name-namespace
However it does not have good support for npm scoped packages yet. There
is a proposal to "fix" this but it has not been implemented yet.
https://github.com/BuckleScript/bucklescript/issues/2405

To "work around" this problem, we can "manually" create the wrappers for
such namespaces by prefixing the file names.
http://2ality.com/2017/12/modules-reasonml.html#namespacing-modules
So in our case I can use a module like `FlopFlip_ReactBroadcast.ConfigureFlopFlip`.